fix(result): 일부 아이템 공용화
